+++ packages/desktop/src/renderer/components/CreateJoinCommunity/JoinCommunity/JoinCommunity.test.tsx
@@ -115,7 +115,7 @@ describe('join community', () => {
     expect(screen.getAllByText('Join with invite link')).toHaveLength(1)
 
     await userEvent.click(screen.getByTestId('paste-a-link'))
-    expect(await screen.findByRole('heading', { name: 'Paste a link to Join', level: 3 })).toBeVisible()
+    expect(await screen.findByRole('heading', { name: 'Paste a link to join', level: 3 })).toBeVisible()
     expect(screen.getByPlaceholderText('Link')).toBeVisible()
 
     await userEvent.click(screen.getByTestId('joinCommunityModalBack'))
@@ -247,7 +247,7 @@ describe('join community', () => {
 
       await userEvent.click(screen.getByTestId('join-with-qr-code'))
       await userEvent.click(await screen.findByTestId('qr-scanner-paste-link'))
-      expect(await screen.findByRole('heading', { name: 'Paste a link to Join', level: 3 })).toBeVisible()
+      expect(await screen.findByRole('heading', { name: 'Paste a link to join', level: 3 })).toBeVisible()
       expect(screen.getByPlaceholderText('Link')).toBeVisible()
 
       await userEvent.click(screen.getByTestId('joinCommunityModalBack'))
@@ -325,7 +325,7 @@ describe('join community', () => {
     const handleCommunityAction = jest.fn()
 
     const result = renderComponent(
-      <PasteLinkComponent heading={'Paste a link to Join'} handleCommunityAction={handleCommunityAction} />,
+      <PasteLinkComponent heading={'Paste a link to join'} handleCommunityAction={handleCommunityAction} />,
       store
     )
 
@@ -350,7 +350,7 @@ describe('join community', () => {
       const handleCommunityAction = jest.fn()
 
       const result = renderComponent(
-        <PasteLinkComponent heading={'Paste a link to Join'} handleCommunityAction={handleCommunityAction} />,
+        <PasteLinkComponent heading={'Paste a link to join'} handleCommunityAction={handleCommunityAction} />,
         store
       )
 
@@ -373,7 +373,7 @@ describe('join community', () => {
     const handleCommunityAction = jest.fn()
 
     const result = renderComponent(
-      <PasteLinkComponent heading={'Paste a link to Join'} handleCommunityAction={handleCommunityAction} />,
+      <PasteLinkComponent heading={'Paste a link to join'} handleCommunityAction={handleCommunityAction} />,
       store
     )
 
@@ -405,7 +405,7 @@ describe('join community', () => {
     const handleCommunityAction = jest.fn()
 
     renderComponent(
-      <PasteLinkComponent heading={'Paste a link to Join'} handleCommunityAction={handleCommunityAction} />,
+      <PasteLinkComponent heading={'Paste a link to join'} handleCommunityAction={handleCommunityAction} />,
       store
     )
 
@@ -427,7 +427,7 @@ describe('join community', () => {
 
     const result = renderComponent(
       <PasteLinkComponent
-        heading={'Paste a link to Join'}
+        heading={'Paste a link to join'}
         handleCommunityAction={handleCommunityAction}
         isConnectionReady={false}
       />,